agregamos localidades por provincia

+++ TP 5 Final/EncomiendaCorrespondencia.cs
@@ -8,7 +8,7 @@ namespace TP_5_Final
 {
     public class EncomiendaCorrespondencia
     {
-        public EncomiendaCorrespondencia(int numeroIdentificacion, float peso)
+        public EncomiendaCorrespondencia(int numeroIdentificacion, decimal peso)
         {
             NumeroIdentificacion = numeroIdentificacion;
             Peso = peso;
@@ -16,7 +16,7 @@ namespace TP_5_Final
 
         public int NumeroIdentificacion { get; set; }
         // public String Descripcion { get; set; }
-        public float Peso { get; set; }
+        public decimal Peso { get; set; }
         public String CartaDePorte { get; set; }
 
     }
